asr_service/whisper_service/requirements.txt
fastapi
loguru
uvicorn
openai-whisper
torch
pydantic